Move bind_info files to table_function subdirectory
- lib/duckdb/table_function_bind_info.rb -> lib/duckdb/table_function/bind_info.rb
- test/duckdb_test/table_function_bind_info_test.rb -> test/duckdb_test/table_function/bind_info_test.rb
- Update require path in lib/duckdb.rb
