+++ lib/duckdb.rb
@@ -20,7 +20,7 @@ require 'duckdb/function_info'
 require 'duckdb/vector'
 require 'duckdb/data_chunk'
 require 'duckdb/table_function'
-require 'duckdb/table_function_bind_info'
+require 'duckdb/table_function/bind_info'
 require 'duckdb/infinity'
 require 'duckdb/instance_cache'
 require 'duckdb/casting'
