+++ lib/duckdb/table_function/bind_info.rb
