+++ test/duckdb_test/table_function/bind_info_test.rb
